feat: added security check to vehicles

+++ PrimeDriveBackend/src/main/java/com/example/PrimeDriveBackend/service/VehicleSpecsService.java
@@ -2,7 +2,10 @@ package com.example.PrimeDriveBackend.service;
 
 import java.util.List;
 
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import com.example.PrimeDriveBackend.repository.VehicleRepository;
+import com.example.PrimeDriveBackend.model.Vehicle;
 
 import com.example.PrimeDriveBackend.Dto.VehicleSpecsDto;
 import com.example.PrimeDriveBackend.Mapper.VehicleSpecsMapper;
@@ -16,6 +19,7 @@ import lombok.RequiredArgsConstructor;
 public class VehicleSpecsService {
     private final VehicleSpecsRepository vehicleSpecsRepository;
     private final VehicleSpecsMapper vehicleSpecsMapper;
+    private final VehicleRepository vehicleRepository;
 
     public List<VehicleSpecsDto> getAllSpecs() {
         return vehicleSpecsRepository.findAll()
@@ -44,6 +48,14 @@ public class VehicleSpecsService {
         VehicleSpecs existing = vehicleSpecsRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Specs not found with id: " + id));
 
+        Vehicle vehicle = vehicleRepository.findBySpecs_Id(id)
+                .orElseThrow(() -> new RuntimeException("No vehicle found for given specs ID"));
+
+        String currentUserId = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (!vehicle.getUsers().equals(currentUserId)) {
+            throw new SecurityException("You are not authorized to update these specs.");
+        }
+
         VehicleSpecs updatedSpecs = vehicleSpecsMapper.toEntity(updatedDto);
         updatedSpecs.setId(existing.getId());
 
@@ -54,6 +66,15 @@ public class VehicleSpecsService {
         if (!vehicleSpecsRepository.existsById(id)) {
             throw new RuntimeException("Specs not found with id: " + id);
         }
+
+        Vehicle vehicle = vehicleRepository.findBySpecs_Id(id)
+                .orElseThrow(() -> new RuntimeException("No vehicle found for given specs ID"));
+
+        String currentUserId = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (!vehicle.getUsers().equals(currentUserId)) {
+            throw new SecurityException("You are not authorized to delete these specs.");
+        }
+
         vehicleSpecsRepository.deleteById(id);
     }
 }
